grapple/extensions.py
# grapple/extensions.py

from flask_sqlalchemy import SQLAlchemy
from flask_migrate import Migrate
from flask_login import LoginManager
from flask_bcrypt import Bcrypt
from flask_wtf.csrf import CSRFProtect
from flask_moment import Moment
from flask_mail import Mail

# Instantiate extensions here. They are not yet attached to the app.
db = SQLAlchemy()
migrate = Migrate()
login_manager = LoginManager()
bcrypt = Bcrypt()
csrf = CSRFProtect()
moment = Moment()
mail = Mail()