requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-Login==0.6.3
Flask-WTF==1.2.2
Flask-Moment==1.0.6
email-validator==2.2.0
PyMySQL==1.0.3
python-dotenv==1.0.0
Werkzeug==2.3.4
gunicorn==20.1.0
Jinja2-Time
flask-bcrypt
wtforms-sqlalchemy